requirements.txt
# requirements.txt
openpyxl>=3.1.5
pandas>=2.3.0
PyQt6>=6.10.0
fpdf2>=2.7.0